- Fix problems about new Yeareducation - Add OnlineExams to students workbooks also on dashboard reports - Add false, true and blank count on OnlineExams Excel report - Add option for Calculate Negative Score in OnlineExam Excel report - Fix title of some colums on OnlineExam excel report - Add option to exams report about not showing course that dosnt have any exam - Add Themplate to attributes list - Add Course, ExamType and Workbook to OnlineExams - Add option that show score for registred item after done

+++ Model/SM_Course.cs
@@ -138,7 +138,7 @@ namespace SCMR_Api.Model
         }
 
 
-        private double getJustifiedScore(double score, int topScore)
+        private double getJustifiedScore(double score, double topScore)
         {
             double mix = (double)20 / topScore;
 
@@ -161,7 +161,7 @@ namespace SCMR_Api.Model
     {
         public double Score { get; set; }
 
-        public int TopScore { get; set; }
+        public double TopScore { get; set; }
 
         public int CourseId { get; set; }
     }
